backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
requests==2.32.5
anthropic
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0
opencensus-ext-azure==1.1.13
python-dotenv==1.0.0
gunicorn==21.2.0
